refactor: Criando molecula para o create menu
